backend/code/test/requirements.txt
pytest
pytest-mock
boto3
moto
aws_lambda_powertools
fastjsonschema
pyyaml